Tests improvements and code bugs fix

+++ db/dbHelper.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 const dbConnection = mongoose.connection;
-const dbDebugger = require('debug')('app:db');
+const dbDebugger = require('debug')('db:helper');
 const dbErr = require('debug')('app:err');
 
 //Needed variables to connect
@@ -9,36 +9,48 @@ const options = {
     useNewUrlParser: true,
     keepAlive: true,
     useUnifiedTopology: true,
-    keepAliveInitialDelay: 300000
+    serverSelectionTimeoutMS: 100,
+    keepAliveInitialDelay: 1000
 };
 
 async function setupConnection(){
     dbDebugger('setupConnection');
     if(dbConnection._readyState != 1){    //Default 0, Connected 'db.once'-> 1
         dbDebugger('Connecting to DB');
-        await mongoose.connect(uri, options, function(err){
-            if(err){
-                dbErr('Could not connect to the DB');
-                return(undefined);
-            }
-        });
-        //db.on((err) => dbErr('Could not connect to DB \n' + err));
-        dbConnection.once(('open'), function(){
-            dbDebugger('Connected to the DB');
-        });
+        let err = await mongoose.connect(uri, options)
+                      .then(() => dbDebugger( 'Connected' ))
+                      .catch( (err) => dbErr("Could not connect "+err)
+        );
     }
+    //db.on((err) => dbErr('Could not connect to DB \n' + err));
+    dbConnection.once(('open'), function(){
+        dbDebugger('Connected to the DB');
+    });
 };
 
 async function getItem (id, model, next){
-    setupConnection();
+    await setupConnection();
+    if(mongoose.connection.readyState != 1)
+        return(false)
     const result = await model.find({_id: id});
-    dbDebugger('Full item retrieved \n' +result)
+    dbDebugger('getItem retrieved \n' +result)
+    next(result);
+}
+
+async function getItemSelectParams (objectToFind, model, params, next){
+    await setupConnection();
+    if(mongoose.connection.readyState != 1)
+        return(false)
+    const result = await model.find(objectToFind).select(params);
+    dbDebugger('Item retrieved \n' +result)
     next(result);
 }
 
 async function getItemPopulated (id, model, fieldToPopulate, 
                                     selectedPopulatedFields="", itemFields="", next){
-    setupConnection();
+    await setupConnection();
+    if(mongoose.connection.readyState != 1)
+        return(false)
     const result = await model.find({_id: id})
                                .populate(fieldToPopulate, " "+selectedPopulatedFields)
                                .select(itemFields);
@@ -48,12 +60,18 @@ async function getItemPopulated (id, model, fieldToPopulate,
 
 async function addItem(params, model, next){
     await setupConnection();
+    if(mongoose.connection.readyState != 1){
+        dbDebugger("addItem - Connection state: "+mongoose.connection.readyState)
+        return(false)
+    }
+    dbDebugger("addItem - connected")
     const newItem = new model(params);
     dbDebugger(newItem);
 
     var result;
     try{
         result = await newItem.save();     //Here is where the validation is performed
+        dbDebugger("addItem retrieved "+result)
         dbDebugger('Added');
     }
     catch(ex){
@@ -68,15 +86,18 @@ async function addItem(params, model, next){
 }
 
 async function updateItem(id, model, params, next){
-    setupConnection();
+    await setupConnection();
+    if(mongoose.connection.readyState != 1)
+        return(false)
     dbDebugger(id);
     //Check it exists
     const item = await model.findById(id);
     if(!item){
-        dbDebugger("Didn't find the diver");
+        dbDebugger("Didn't find the item");
         next(undefined);
         return;
     }
+    dbDebugger("updateItem retrieved "+item)
     for (var field in model.schema.paths) {
         if ((field !== '_id') && (field !== '__v')) {
            if (params[field] !== undefined) {
@@ -89,8 +110,11 @@ async function updateItem(id, model, params, next){
 };
 
 async function deleteItem(id, model, next) {
-    setupConnection();
+    let connection = await setupConnection();
+    if(connection == false)
+        return(false)
     const result = await model.deleteMany({_id: id});
+    dbDebugger("deleteItem retrieved "+result)
     next(result);
 };
 
@@ -101,3 +125,4 @@ module.exports.addItem = addItem;
 module.exports.connection = dbConnection;
 module.exports.updateItem = updateItem;
 module.exports.deleteItem = deleteItem;
+module.exports.getItemSelectParams = getItemSelectParams;
